Rename irrelevant identifiers and refactor models
