--- usg-cekirge-challenge/Constants/Constant.swift
@@ -1,12 +1,0 @@
-//
-//  Constant.swift
-//  usg-cekirge-challenge
-//
-//  Created by Hakan Or on 20.05.2023.
-//
-
-import Foundation
-
-final class Constant {
-    
-}
