+++ usg-cekirge-challenge/Model/Location.swift
@@ -16,10 +16,3 @@ struct Location: Codable {
     let url: String
     let created: String
 }
-
-struct Info: Codable {
-    let count: Int
-    let pages: Int
-    let next: String?
-    let prev: String?
-}
