+++ usg-cekirge-challenge/Scene/CharacterDetails/CharacterDetailsViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  CharacterDetailsViewModel.swift
+//  usg-cekirge-challenge
+//
+//  Created by Hakan Or on 6.05.2024.
+//
+
+import Foundation
